Refactor more `PlotView`
* Add API type
* Extract utility function
* Remove unneeded code
